Location Format, Default Tribes, Help Text
Added a formatter to the Location class to make the X and Y coordinates
print out better.

Added a number of static values for default tribe generation to the
TribeManager class, with intent to make these configurable in the
future.  Added the addDefaultTribe() method to auto-generate a tribe
using these default values. Added the isTooClose() method to check if a
given tribe, whether generated automatically or manually, is too close
to any existing ones.

Added a new help command, to print out the existing commands to choose
from. Added a "default" subcommand to the `make tribe` command to fill
out up to the maximum with default details for new tribes. Updated all
makeTribe() logic to use the new isTooClose() logic.

Updated some help messaging to be more professional.