Printable, PrintManager
Created PrintManager class for managing print interactions -- whether to
console or GUI. The default print operation is to use the system print
methods.

Created Printable trait. This creates static print() and println()
methods that implementing classes will now automatically be using
instead of the system print methods. Additionally added the DBG() method
to it and removed from the other classes that were implementing it
manually.