GUI Furthering
Expanded on the GUI: added a tabbed pane with dummy data, added a text
field for entering information to be read by the simulation. Hitting
enter sends the code to the simulation and clears the field.

Modified the Simulation class's logic to operate differently when the
boolean USE_GUI is set to true, so that it operates entirely differently
when using the GUI.

Added *magic* to fix the problem with reading from console. See code for
the magic.