Scrollable, Tabbable, Debuggable, Commandable
Put the console into a JScrollPane and set the size correctly. Adjusted
the scaling of the MigLayout components to stop overflow of the
JTabbedPane.

Added dummy data for the JTabbedPane to showcase our intended future
vision.

Added a safety measure to ensure new commands can't be run while others
are underway. Additionally, made it so the console can't be directly
edited by the user.