Info Command
Added info command, as well as info() method skeleton. Not the same as
the info() method in the Informable interface. Calls the info() method
ont he specified object and prints it in an organized way.

Needs to be fleshed out. TODOs in the info() method here lay out what
needs to happen.