Console Additions
Replaced font in console with Consolas, a newer font.

Changed behavior of down arrow in text field to no longer beep() at the
user, and instead save the previously-entered text to history and clear.
Additionally set the GUI to request focus on that text field on start.

Added a variable for user prefix, to now print the commands that people
run to the console.