Stuff missed 2023-05-31
A few changes I forgot to commit.

Suggestion box -- future UI update to show the user possible commands
they could complete.

Tests -- a script testing the DSL we are considering for configuring
threats and the like.

Manifest.

Simulation class -- a single TODO reminding us of the Command class we
need to make to support the suggestion box.

ScriptManager class -- a new class implementing the tests in the
affects.groovy file.